requirements.txt
psycopg2
django
djangorestframework
django-redis

pytest
pytest-django

drf-spectacular
